# website/static/json/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "couic",
        "description": "A lightweight eBPF-powered network filtering solution controllable through a REST API",
        "license": {
            "name": "MIT OR Apache-2.0",
            "identifier": "MIT OR Apache-2.0"
        },
        "version": "0.5.12"
    },
...
    "paths": {
...
        "/v1/{policy}": {
            "get": {
                "tags": [
                    "policies"
                ],
                "summary": "List all entries based on policy",
                "operationId": "list_entries",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (e.g., 'drop', 'ignore')",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Entry"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "policies"
                ],
                "summary": "Create a new entry",
                "description": "Add a new CIDR entry to the specified policy (drop or ignore).",
                "operationId": "create_entry",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (e.g., 'drop', 'ignore')",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "ignore"
                    }
                ],
                "requestBody": {
                    "description": "Entry to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EntryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Entry created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (malformed JSON or invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v1/{policy}/peer": {
            "post": {
                "tags": [
                    "policies"
                ],
                "summary": "Synchronize entries with peers",
                "description": "Process a batch of peer synchronization jobs (add or remove entries).",
                "operationId": "peer_entries",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        }
                    }
                ],
                "requestBody": {
                    "description": "List of peer jobs to process",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PeerJob"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Peer jobs processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PeerJob"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (malformed JSON or invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v1/{policy}/{ip}/{prefix}": {
            "get": {
                "tags": [
                    "policies"
                ],
                "summary": "Get a specific entry based on policy",
                "operationId": "get_entry",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        }
                    },
                    {
                        "name": "ip",
                        "in": "path",
                        "description": "IP address (IPv4 or IPv6)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "examples": [
                                "192.168.1.0",
                                "2001:db8::"
                            ]
                        }
                    },
                    {
                        "name": "prefix",
                        "in": "path",
                        "description": "CIDR prefix length",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 128,
                            "examples": [
                                24,
                                64
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Entry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "policies"
                ],
                "summary": "Delete an entry based on policy",
                "operationId": "delete_entry",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        }
                    },
                    {
                        "name": "ip",
                        "in": "path",
                        "description": "IP address (IPv4 or IPv6)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "examples": [
                                "192.168.1.0",
                                "2001:db8::"
                            ]
                        }
                    },
                    {
                        "name": "prefix",
                        "in": "path",
                        "description": "CIDR prefix length",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 128,
                            "examples": [
                                24,
                                64
                            ]
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Entry deleted"
                    },
                    "400": {
                        "description": "Bad request (invalid path parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompositeError": {
                "type": "object",
                "required": [
                    "code",
                    "message",
                    "errors"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "Entry": {
                "type": "object",
                "description": "Represents a firewall entry with creation timestamp (returned by API)",
                "required": [
                    "creation",
                    "cidr",
                    "expiration"
                ],
                "properties": {
                    "creation": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Unix timestamp when the entry was created",
                        "examples": [
                            1680000000
                        ]
                    },
                    "cidr": {
                        "type": "string",
                        "description": "CIDR notation (IPv4 or IPv6)",
                        "examples": [
                            "192.168.1.0/24"
                        ]
                    },
                    "expiration": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Unix timestamp when the entry expires (0 = never)",
                        "examples": [
                            1680000000
                        ]
                    },
                    "tag": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Optional tag for categorization",
                        "maxLength": 64,
                        "pattern": "^[a-zA-Z0-9._-]*$",
                        "examples": [
                            "malicious"
                        ]
                    }
                }
            },
            "EntryRequest": {
                "type": "object",
                "description": "Request body for creating a firewall entry",
                "required": [
                    "cidr",
                    "expiration"
                ],
                "properties": {
                    "cidr": {
                        "type": "string",
                        "description": "CIDR notation (IPv4 or IPv6)",
                        "examples": [
                            "192.168.1.0/24"
                        ]
                    },
                    "expiration": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Unix timestamp when the entry expires (0 = never)",
                        "examples": [
                            1680000000
                        ]
                    },
                    "tag": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Optional tag for categorization",
                        "maxLength": 64,
                        "pattern": "^[a-zA-Z0-9._-]*$",
                        "examples": [
                            "malicious"
                        ]
                    },
                    "metadata": {
                        "$ref": "#/components/schemas/Metadata",
                        "description": "Optional metadata for the entry"
                    }
                }
            },
            "Metadata": {
                "type": "object",
                "description": "Optional metadata for an entry",
                "properties": {
                    "kind": {
                        "type": "string",
                        "description": "Type or category of the metadata"
                    },
                    "detail": {
                        "type": "string",
                        "description": "Detailed description"
                    },
                    "extra": {
                        "type": "object",
                        "description": "Additional arbitrary metadata as key-value pairs",
                        "additionalProperties": true
                    }
                }
            },
            "ErrorCode": {
                "type": "string",
                "description": "Enum representing error codes used in the API",
                "enum": [
                    "Eprocessing",
                    "Eunauthorized",
                    "Enotfound",
                    "Econflict",
                    "Ebadrequest",
                    "Einvalid",
                    "Einternal",
                    "Enotimplemented"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "PeerJob": {
                "type": "object",
                "description": "A peer synchronization job",
                "required": [
                    "action",
                    "entry"
                ],
                "properties": {
                    "action": {
                        "$ref": "#/components/schemas/Action"
                    },
                    "entry": {
                        "$ref": "#/components/schemas/EntryRequest"
                    }
                }
            },
            "Action": {
                "type": "string",
                "description": "Action to perform on an entry",
                "enum": [
                    "add",
                    "remove"
                ]
            },
...
            "Policy": {
                "type": "string",
                "description": "Firewall policy type",
                "enum": [
                    "drop",
                    "ignore"
                ]
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Uuidv4",
                "description": "An API key is a token that you provide when making API calls. Include the token in a header parameter called `Authorization`.\nExample: `Authorization: Bearer b03dd133-3827-4793-b0df-37f6ffbdd9a9`"
            }
        }
    },
...
}